Rearrange to group `proto/` together, and remove `rust_analyzer/`
