+++ crates/lsp/src/diff.rs
