--- crates/lsp/src/rust_analyzer/mod.rs
@@ -1,5 +1,0 @@
-pub mod diff;
-pub mod line_index;
-pub mod text_edit;
-pub mod to_proto;
-pub mod utils;
