--- crates/lsp/src/rust_analyzer/utils.rs
@@ -1,56 +1,0 @@
-// --- source
-// authors = ["rust-analyzer team"]
-// license = "MIT OR Apache-2.0"
-// origin = "https://github.com/rust-lang/rust-analyzer/blob/master/crates/rust-analyzer/src/lsp/utils.rs"
-// ---
-
-use std::ops::Range;
-
-use tower_lsp::lsp_types;
-
-use crate::from_proto;
-use crate::proto::PositionEncoding;
-
-pub(crate) fn apply_document_changes(
-    encoding: PositionEncoding,
-    file_contents: &str,
-    mut content_changes: Vec<lsp_types::TextDocumentContentChangeEvent>,
-) -> String {
-    // If at least one of the changes is a full document change, use the last
-    // of them as the starting point and ignore all previous changes.
-    let (mut text, content_changes) = match content_changes
-        .iter()
-        .rposition(|change| change.range.is_none())
-    {
-        Some(idx) => {
-            let text = std::mem::take(&mut content_changes[idx].text);
-            (text, &content_changes[idx + 1..])
-        }
-        None => (file_contents.to_owned(), &content_changes[..]),
-    };
-    if content_changes.is_empty() {
-        return text;
-    }
-
-    let mut line_index = biome_line_index::LineIndex::new(&text);
-
-    // The changes we got must be applied sequentially, but can cross lines so we
-    // have to keep our line index updated.
-    // Some clients (e.g. Code) sort the ranges in reverse. As an optimization, we
-    // remember the last valid line in the index and only rebuild it if needed.
-    // The VFS will normalize the end of lines to `\n`.
-    let mut index_valid = !0u32;
-    for change in content_changes {
-        // The None case can't happen as we have handled it above already
-        if let Some(range) = change.range {
-            if index_valid <= range.end.line {
-                line_index = biome_line_index::LineIndex::new(&text);
-            }
-            index_valid = range.start.line;
-            if let Ok(range) = from_proto::text_range(&line_index, range, encoding) {
-                text.replace_range(Range::<usize>::from(range), &change.text);
-            }
-        }
-    }
-    text
-}
