+++ crates/lsp/src/test/client_ext.rs
@@ -2,7 +2,8 @@ use biome_text_size::TextRange;
 use lsp_test::lsp_client::TestClient;
 use tower_lsp::lsp_types;
 
-use crate::{documents::Document, from_proto, to_proto};
+use crate::documents::Document;
+use crate::proto::{from_proto, to_proto};
 
 pub(crate) trait TestClientExt {
     async fn open_document(
@@ -66,7 +67,7 @@ impl TestClientExt for TestClient {
 
     async fn format_document(&mut self, doc: &Document, filename: FileName) -> String {
         match self.format_document_edits(doc, filename).await {
-            Some(edits) => from_proto::apply_text_edits(doc, edits).unwrap(),
+            Some(edits) => apply_text_edits(doc, edits).unwrap(),
             None => doc.contents.clone(),
         }
     }
@@ -80,7 +81,7 @@ impl TestClientExt for TestClient {
         let Some(edits) = self.format_document_range_edits(doc, filename, range).await else {
             return doc.contents.clone();
         };
-        from_proto::apply_text_edits(doc, edits).unwrap()
+        apply_text_edits(doc, edits).unwrap()
     }
 
     async fn format_document_edits(
@@ -121,7 +122,7 @@ impl TestClientExt for TestClient {
     ) -> Option<Vec<lsp_types::TextEdit>> {
         let lsp_doc = self.open_document(doc, filename).await;
 
-        let range = to_proto::range(&doc.line_index, range, doc.line_index.encoding).unwrap();
+        let range = to_proto::range(range, &doc.line_index.index, doc.position_encoding).unwrap();
 
         self.range_formatting(lsp_types::DocumentRangeFormattingParams {
             text_document: lsp_types::TextDocumentIdentifier {
@@ -157,4 +158,23 @@ fn formatting_options(doc: &Document) -> lsp_types::FormattingOptions {
         insert_spaces: matches!(indent_style, settings::IndentStyle::Space),
         ..Default::default()
     }
+}
+
+fn apply_text_edits(doc: &Document, mut edits: Vec<lsp_types::TextEdit>) -> anyhow::Result<String> {
+    let mut text = doc.contents.clone();
+
+    // Apply edits from bottom to top to avoid inserted newlines to invalidate
+    // positions in earlier parts of the doc (they are sent in reading order
+    // accorder to the LSP protocol)
+    edits.reverse();
+
+    for edit in edits {
+        let range =
+            from_proto::text_range(edit.range, &doc.line_index.index, doc.position_encoding)?;
+        let start: usize = range.start().into();
+        let end: usize = range.end().into();
+        text.replace_range(start..end, &edit.new_text);
+    }
+
+    Ok(text)
 }
